[Keymap] User keymap for AKB/Raine (#8081)
* Adding my keymap

Adding a personalized keymap to the Raine keyboard. This diffirentiates itself by being useble with nordic characters even when the bottom row has blockers. Also added ASCII art to help visualize what you are doing when programming the board.

* Update Readme.md

Ned image of the board

* Update Readme.md

Forgot a space in Readme

* Update Readme.md

And another missed space

* Update keymap.c

Renoved Definetion that is not used

* Update Readme.md

* Remove files for renaming

* Adding files after renaming

Was unable to get Girthub to push renamed files/folders. Therefore removeing and adding them again.